Mobile fixes
Improve mobile card navigation by requiring a certain distance to travel before removing the card. This fixes the the issue where just swiping up or down would remove a card
Removing onMouseHover for now because it is causing mobile users to have to tap twice in order to interact with the element.